Remove namespace from models to avoid collision with gem
Renamed ActiveIntelligence::Conversation -> Conversation
Renamed ActiveIntelligence::Message -> Message

This eliminates the namespace collision between the gem's ActiveIntelligence
module and the Rails models. Much cleaner and simpler approach.

Changes:
- Moved models out of active_intelligence directory
- Removed module wrapper from model classes
- Updated all controller references
- Removed namespace module file
- Removed explicit loading initializer
- Updated application config
