--- examples/rails_bible_chat/app/models/active_intelligence.rb
@@ -1,4 +1,0 @@
-# Namespace for ActiveIntelligence models
-# This file ensures Rails autoloading works correctly for namespaced models
-module ActiveIntelligence
-end
